fix shape64

+++ ppocr/modeling/heads/rec_sar_head.py
@@ -242,14 +242,14 @@ class ParallelSARDecoder(BaseDecoder):
         # bsz * (seq_len + 1) * h * w * attn_size
         attn_weight = self.conv1x1_2(attn_weight)
         # bsz * (seq_len + 1) * h * w * 1
-        bsz, T, h, w, c = paddle.shape(attn_weight).astype("int32")
+        bsz, T, h, w, c = paddle.shape(attn_weight)
         assert c == 1
 
         if valid_ratios is not None:
             # cal mask of attention weight
             for i in range(valid_ratios.shape[0]):
                 valid_width = paddle.minimum(
-                    w, paddle.ceil(valid_ratios[i] * w).astype("int32")
+                    w, paddle.ceil(valid_ratios[i] * w).astype("int64")
                 )
                 if valid_width < w:
                     attn_weight[i, :, :, valid_width:, :] = float("-inf")
